Improve update handling
ObjectId in the Esri is not stable between days. Use the date of the most recent report to see if there's anything new to fetch.

Records may be updated, so if there is new data available, we re-fetch everything.

Page cache (invalidation) is made a bit more intelligent.

The data process tasks only checks days with modified records.